Hestabit_Assignmrnt/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
dj-database-url>=2.2
Jinja2>=3.1
